backend/requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
PyYAML==6.0.1
Requests==2.31.0
